move generation of old records to an extension
