+++ tests/php/Task/FluentMigrationTaskTest/TranslatedDataObject.php
@@ -4,12 +4,8 @@
 namespace TractorCow\Fluent\Tests\Task\FluentMigrationTaskTest;
 
 
-use SilverStripe\Core\ClassInfo;
-use SilverStripe\Core\Config\Config;
 use SilverStripe\Dev\TestOnly;
 use SilverStripe\ORM\DataObject;
-use SilverStripe\ORM\DB;
-use SilverStripe\ORM\Queries\SQLUpdate;
 use TractorCow\Fluent\Extension\FluentExtension;
 
 /**
@@ -24,7 +20,8 @@ use TractorCow\Fluent\Extension\FluentExtension;
 class TranslatedDataObject extends DataObject implements TestOnly
 {
     private static $extensions = [
-        FluentExtension::class
+        FluentExtension::class,
+        OldFluentDataExtension::class
     ];
 
     private static $db = [
@@ -49,46 +46,4 @@ class TranslatedDataObject extends DataObject implements TestOnly
         'Name_de_AT'
     ];
 
-    public function onAfterWrite()
-    {
-        $ancestry = array_reverse(ClassInfo::ancestry(static::class));
-        $schema = static::getSchema();
-
-        foreach ($ancestry as $class) {
-            if (!$schema->classHasTable($class)) {
-                continue;
-            }
-            $table = $schema->tableName($class);
-            $update = SQLUpdate::create()
-                ->setTable($table)
-                ->setWhere('ID = ' . $this->ID);
-
-            foreach ($class::config()->get('old_fluent_fields', Config::UNINHERITED) as $field) {
-                $update->assign($field, $this->getField($field)); //value is set from fixtures in $this->record
-                $update->execute();
-            }
-
-        }
-        parent::onAfterWrite();
-    }
-
-
-    /**
-     * Helper to get the old *_translationgroups table and the Locale field created
-     */
-    public function requireTable()
-    {
-        parent::requireTable();
-
-//        $baseDataClass = DataObject::getSchema()->baseDataClass($this->ClassName);
-//        if ($this->ClassName != $baseDataClass) {
-//            return;
-//        }
-
-        $schemaManager = DB::get_schema();
-
-        foreach (static::config()->get('old_fluent_fields') as $field) {
-            $schemaManager->requireField(self::getSchema()->tableName(static::class), $field, 'VARCHAR(255) NULL DEFAULT NULL ');
-        }
-    }
 }
